timob-1512: addressed comments
